requirements.txt
# ── Satellite & GEE ─────────────────────────────────────────────────────────
earthengine-api>=0.1.380
google-auth>=2.0
numpy>=1.24
Pillow>=10.0
tifffile>=2023.1

# ── Telegram bot ─────────────────────────────────────────────────────────────
python-telegram-bot>=20.7
apscheduler>=3.10

# ── PDF reports ──────────────────────────────────────────────────────────────
fpdf2>=2.7

# ── FastAPI web server ───────────────────────────────────────────────────────
fastapi>=0.110
uvicorn[standard]>=0.29
pydantic>=2.0

# ── Google OAuth SSO ─────────────────────────────────────────────────────────
authlib>=1.3
httpx>=0.27          # authlib async HTTP client

# ── JWT (cookies) ────────────────────────────────────────────────────────────
python-jose[cryptography]>=3.3

# ── Razorpay payments ────────────────────────────────────────────────────────
razorpay>=1.4

# ── Celery + Redis (background workers) ──────────────────────────────────────
celery>=5.3
redis>=5.0

# ── Cloudflare R2 (S3-compatible image store) ────────────────────────────────
boto3>=1.34

# ── PostgreSQL (Railway production) ──────────────────────────────────────────
psycopg2-binary>=2.9

# ── Email (Resend API) ───────────────────────────────────────────────────────
resend>=0.7

# ── Shared utilities ─────────────────────────────────────────────────────────
requests>=2.31
python-dotenv>=1.0
pytz
